requirements.txt
absl-py==2.1.0
accelerate==1.5.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1732732445543/work
annotated-types @ file:///croot/annotated-types_1709542908624/work
anyio==4.8.0
archspec @ file:///croot/archspec_1709217642129/work
attrs==25.3.0
boltons @ file:///croot/boltons_1737061692168/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1738623731865/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
chex==0.1.89
click==8.1.8
conda @ file:///croot/conda_1738168370816/work
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1736524596117/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1739298990519/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1737733694612/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1731366181659/work
contourpy==1.3.1
cryptography @ file:///croot/cryptography_1732129644106/work
cycler==0.12.1
datasets==3.3.2
-e git+https://github.com/huggingface/diffusers@82188cef0487837b8c70fc3f36ea63c05c85f341#egg=diffusers
dill==0.3.8
distro @ file:///croot/distro_1714488253808/work
docker-pycreds==0.4.0
einops==0.8.1
etils==1.12.2
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.17.0
flash_attn==2.7.4.post1
flax==0.10.4
fonttools==4.56.0
frozendict @ file:///croot/frozendict_1713194832637/work
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gradio==3.23.0
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanize==4.12.1
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.6.1
importlib_resources==6.5.2
jax==0.5.2
jaxlib==0.5.1
Jinja2==3.1.6
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
libmambapy @ file:///croot/mamba-split_1734469461757/work/libmambapy
lightning-utilities==0.14.3
linkify-it-py==2.0.3
Markdown==3.7
markdown-it-py @ file:///work/perseverance-python-buildout/croot/markdown-it-py_1698846045803/work
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.1
mdit-py-plugins==0.3.3
mdurl @ file:///work/perseverance-python-buildout/croot/mdurl_1698845653285/work
menuinst @ file:///croot/menuinst_1738943416351/work
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.30.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
open_clip_torch==2.31.0
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.8
orjson==3.10.15
packaging @ file:///croot/packaging_1734472117206/work
pandas==2.2.3
peft==0.7.0
pillow==11.1.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///croot/pluggy_1733169602837/work
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.1
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///croot/pydantic_1734736067156/work
pydantic_core @ file:///croot/pydantic-core_1734726052986/work
pydub==0.25.1
Pygments @ file:///work/perseverance-python-buildout/croot/pygments_1698846270603/work
pyparsing==3.2.1
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
rich @ file:///croot/rich_1732638981168/work
rpds-py==0.23.1
ruamel.yaml @ file:///croot/ruamel.yaml_1727980156079/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1727769819935/work
safetensors==0.5.3
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
setuptools==75.8.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorstore==0.1.72
tf==1.0.3
timm==1.0.15
tokenizers==0.21.1
toolz==1.0.0
torch==2.6.0
torchmetrics==1.7.1
torchvision==0.21.0
tqdm @ file:///croot/tqdm_1738943501192/work
transformers==4.49.0
treescope==0.1.9
triton==3.2.0
truststore @ file:///croot/truststore_1736550121485/work
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.1
uc-micro-py==1.0.3
urllib3 @ file:///croot/urllib3_1737133630106/work
uvicorn==0.34.0
wandb==0.19.8
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard @ file:///croot/zstandard_1731356346222/work
